requirements.txt
fastapi
uvicorn
torch
torchvision
python-multipart
albumentations
pillow
pycocotools
pytest
requests
